requirements.txt
# Core
langchain==0.1.15
openai==1.30.1
twilio==8.11.0
sendgrid==6.11.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
langgraph
pandas

# Optional: for memory and vector DB
chromadb==0.4.24
sentence-transformers==2.5.1

# Utility
python-dotenv==1.0.1
requests==2.31.0
PyYAML==6.0.1
